datapath: Decouple IPV4_MASQUERADE from IPV4_NODEPORT
Previously, the same IPv4 addr was used for SNAT-ing packets sent from
local endpoints to outside (BPF-masq) and requests sent from outside
clients to a NodePort service to an intermediate node.

The commit [1] made the IPV4_NODEPORT selection to be based on k8s
NodeIP.  I.e. if a device has multiple IPv4 addrs, then the one which is
used as a k8s NodeIP is preferred. This change might have broken
environments in which a single device has two IP addrs: private (k8s
NodeIP) used to communicate among k8s nodes and public - with outside.
In this scenario, packets sent from local endpoints to outside are
SNAT-ed to the private IP addr which is obviously wrong. An example of
such environment is packet.net in "Layer3 networking mode" in which two
NICs are bonded and have public and private IP addrs assigned.

To fix this, introduce IPV4_MASQUERADE which is decoupled from
IPV4_NODEPORT, and is determined by the same "firstGlobalV4Addr" routine
as the latter, but with the preference for public IP addrs.

[1]: 870c02c3 ("Makes k8sNodeIP the preferred IP when initializing NodePort addresses".)

Signed-off-by: Martynas Pumputis <m@lambda.lt>