doc: simplify Hubble installation instructions
Now that Hubble is enabled by default, Hubble installation instructions
can be simplified to only instruct about enabling Hubble Relay for
cluster-wide visibility. Details about the Hubble architecture are now
omitted to ensure the instructions are easy to follow. A new link that
points to the observability section of the documentation is added. In
this section, users will find details about Hubble architecture should
they be interested.

Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>