pkg/service: Add backends as managed neighbor entry
In LB-only mode, push backends as managed neighbors into the Linux kernel's
neighboring subsystem. This is needed in particular for XDP since in XDP
layer it is not possible to resolve backend L2 addresses if they are in the
same L2 domain. For CNI mode this is not an issue since we push down all
cluster nodes in the same L2 as managed neighbors anyway. However, in the
L4LB case, backend nodes are not in our LB-only cluster. Reuse the same
internal infrastructure for managing and pushing down neighbors, so that
we do not need to open-code it and improvements benefit both L4LB and CNI
mode.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>