l4lb: fix inability to properly update service
When cilium-agent is restarted the restored backends are put into
`restoredBackendHashes` set so that obsolete ones can be properly
deleted only when k8s sync is done.

This is valid in k8s case, but when cilium-agent is running in
standalone L4LB mode there is no way to do sync. There are only the
ServiceAPI calls.
Therefore with the current code it is not possible to restart l4lb
and remove specific backend from a service without previous service
update call (to let l4lb know all the backends - simulate the sync).
This should not be needed.

This commit fixes it in l4lb case.

Signed-off-by: Ondrej Blazek <ondrej.blazek@firma.seznam.cz>